all requirements done

+++ week-6-lab/week-6-lab/EmployeeDatabase.h
@@ -15,11 +15,11 @@
 
 -(NSInteger)count;
 -(NSArray *)allEmployees;
--(Employee *)employeeAtIndex:(int)index;
+-(Employee *)employeeAtIndex:(NSInteger)index;
 
 -(void)add:(Employee *)employee;
 -(void)remove:(Employee *)employee;
--(void)removeEmployeeAtIndex:(int)index;
+-(void)removeEmployeeAtIndex:(NSInteger)index;
 -(void)removeAllEmployees;
 
 @end
